Update dashboard (#108)
Update dashboard + address

+++ app/Http/Controllers/User/DashboardController.php
@@ -12,8 +12,14 @@ use Illuminate\Support\Facades\Hash;
 
 class DashboardController extends Controller
 {
+    //Trang dashboard người dùng
+    public function dashboard()
+    {
+        return view('user.dashboard');
+    }
     public function editProfile(Request $request)
     {
+        /** @var \App\Models\User $user */
         $user = Auth::user();
         $validatedData = $request->validate([
             'full_name' => 'required|string|max:255',
@@ -23,7 +29,6 @@ class DashboardController extends Controller
                 Rule::unique('users', 'email')->ignore($user->id),
             ],
             'phone' => 'required|regex:/^0\d{9,10}$/',  // Số điện thoại bắt đầu bằng 0 và có 10 hoặc 11 chữ số
-            'address' => 'nullable|string',
         ], [
             'full_name.required' => 'Please enter your full name.',
             'phone.required' => 'Please enter phone number.',
@@ -36,13 +41,12 @@ class DashboardController extends Controller
             'full_name' => $validatedData['full_name'],
             'phone' => $validatedData['phone'],
             'email' => $validatedData['email'],
-            'address' => $validatedData['address'],
         ]);
 
         return response()->json([
             'success' => true,
             'message' => 'Updated personal information successfully!',
-            'data' => $user // Trả về đối tượng người dùng đã được cập nhật
+            'data' => $user
         ]);
     }
 
@@ -57,7 +61,7 @@ class DashboardController extends Controller
             'new_password.min' => 'The new password must have at least 6 characters.',
             'new_password.confirmed' => 'Confirm passwords do not match.',
         ]);
-
+        /** @var \App\Models\User $user */
         $user = Auth::user();
 
         // Kiểm tra mật khẩu hiện tại có khớp không
@@ -81,20 +85,22 @@ class DashboardController extends Controller
             'full_name' => 'required|string|max:255',
             'phone_number' => 'required|regex:/^0\d{9,10}$/',
             'address' => 'required|string|max:255',
-        ],[
+        ], [
             'full_name.required' => 'Please enter your full name.',
             'phone_number.required' => 'Please enter your phone number.',
             'phone_number.regex' => 'The phone number must start with 0 and have 10 or 11 digits.',
             'address.required' => 'Please enter your address.'
         ]);
-
+        $hasDefaultAddress = User_shipping_address::where('user_id', auth()->id())
+            ->where('is_active', 1)
+            ->exists();
         // Tạo địa chỉ mới
         $address_shipping = new User_shipping_address;
         $address_shipping->full_name = $request->full_name;
         $address_shipping->phone_number = $request->phone_number;
         $address_shipping->address = $request->address;
         $address_shipping->user_id = auth()->id();
-        $address_shipping->is_active = 0;
+        $address_shipping->is_active = $hasDefaultAddress ? 0 : 1;
         $address_shipping->save();
 
         return response()->json([
@@ -106,49 +112,58 @@ class DashboardController extends Controller
 
     public function editAddress(Request $request, $id)
     {
+        $userId = auth()->user()->id;
         $request->validate([
             'full_name' => 'required|string|max:255',
             'phone_number' => 'required|regex:/^0\d{9,10}$/',
             'address' => 'required|string|max:255',
-        ],[
+        ], [
             'full_name.required' => 'Please enter your full name.',
             'phone_number.required' => 'Please enter your phone number.',
             'phone_number.regex' => 'The phone number must start with 0 and have 10 or 11 digits.',
             'address.required' => 'Please enter your address.'
         ]);
 
-        $address = Auth::user()->user_shipping_addresses()->findOrFail($id);
+        $user = User::findOrFail($userId);
+        $address = $user->user_shipping_addresses()->findOrFail($id);
         $address->update($request->all());
         return response()->json([
             'success' => true,
-            'message' => 'Address updated successfully.'
+            'message' => 'Address updated successfully.',
+            'data' => $address
         ]);
     }
 
     public function deleteAddress($id)
     {
-        // Tìm địa chỉ bằng cách sử dụng where và id của địa chỉ
-        $address = Auth::user()->user_shipping_addresses()->where('id', $id)->first();
+        $userId = auth()->user()->id;
+        $user = User::findOrFail($userId);
+        // Tìm địa chỉ theo ID
+        $address = $user->user_shipping_addresses()->findOrFail($id);
+        // Kiểm tra nếu địa chỉ là mặc định
+        if ($address->is_active == 1) {
+            return back()->with('statusError', 'Bạn không thể xoá địa chỉ mặc định.');
+        }
         $address->delete();
         return back()->with('statusSuccess', 'Địa chỉ đã được xóa thành công.');
     }
 
-    // public function setDefaultShippingAddress($id)
-    // {
-    //     // Lấy tất cả các địa chỉ của người dùng
-    //     $addresses = Auth::user()->user_shipping_addresses;
+    public function setDefaultShippingAddress($id)
+    {
+        // Lấy tất cả các địa chỉ của người dùng
+        $addresses = Auth::user()->user_shipping_addresses;
 
-    //     // Đặt tất cả các địa chỉ thành không phải mặc định
-    //     foreach ($addresses as $address) {
-    //         $address->is_active = 0;
-    //         $address->save();
-    //     }
+        // Đặt tất cả các địa chỉ thành không phải mặc định
+        foreach ($addresses as $address) {
+            $address->is_active = 0;
+            $address->save();
+        }
 
-    //     // Đặt địa chỉ được chọn thành mặc định
-    //     $defaultAddress = User_shipping_address::findOrFail($id);
-    //     $defaultAddress->is_active = 1;
-    //     $defaultAddress->save();
+        // Đặt địa chỉ được chọn thành mặc định
+        $defaultAddress = User_shipping_address::findOrFail($id);
+        $defaultAddress->is_active = 1;
+        $defaultAddress->save();
 
-    //     return redirect()->back()->with('success', 'Địa chỉ mặc định đã được cập nhật thành công.');
-    // }
+        return redirect()->back()->with('statusSuccess', 'Địa chỉ mặc định đã được cập nhật thành công.');
+    }
 }
